# docs/swagger.yaml
definitions:
  model.CartItem:
    properties:
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  model.Order:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.CartItem'
        type: array
      phone:
        type: string
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  model.Product:
    properties:
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      sizes:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /api/cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CartItem'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все товары в корзине
      tags:
      - cart
  /api/cart/{id}:
    get:
      parameters:
      - description: CartItem Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CartItem'
        "404":
          description: cart item not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить товар корзины по ID
      tags:
      - cart
  /api/orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Order'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все заказы
      tags:
      - orders
  /api/orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "404":
          description: order not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить заказ по ID
      tags:
      - orders
  /api/products:
    get:
      description: Возвращает список всех продуктов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все продукты
      tags:
      - products
  /api/products/{id}:
    get:
      description: Возвращает продукт по ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "404":
          description: product not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить продукт по ID
      tags:
      - products
swagger: "2.0"